requirements_app.txt
Flask
psycopg2-binary
Flask-WTF
